migrate and reuse the httpc codes for Http.dispatch/1 behaviour callback default implementation
