+++ test/http_test.exs
@@ -1,9 +1,6 @@
-defmodule Hui.Http.HttpcTest do
+defmodule Hui.HttpTest do
   use ExUnit.Case, async: true
-
   import ExUnit.CaptureLog
-
-  alias Hui.Http.Httpc
   alias Hui.Http
 
   setup do
@@ -19,7 +16,7 @@ defmodule Hui.Http.HttpcTest do
         Plug.Conn.resp(conn, 200, "getting a response")
       end)
 
-      {_, resp} = %Http{url: url} |> Httpc.dispatch()
+      {_, resp} = %Http{url: url} |> Http.dispatch()
 
       assert resp.status == 200
       assert resp.body == "getting a response"
@@ -33,7 +30,7 @@ defmodule Hui.Http.HttpcTest do
         |> Plug.Conn.resp(200, Jason.encode!(json_resp))
       end)
 
-      {_, resp} = %Http{url: url} |> Httpc.dispatch()
+      {_, resp} = %Http{url: url} |> Http.dispatch()
 
       assert resp.body == json_resp
     end
@@ -44,7 +41,7 @@ defmodule Hui.Http.HttpcTest do
         |> Plug.Conn.resp(200, "non json response")
       end)
 
-      {_, resp} = %Http{url: url} |> Httpc.dispatch()
+      {_, resp} = %Http{url: url} |> Http.dispatch()
       assert resp.body == "non json response"
     end
 
@@ -56,7 +53,7 @@ defmodule Hui.Http.HttpcTest do
         Plug.Conn.resp(conn, 200, "getting a response")
       end)
 
-      {_, resp} = %Http{url: url, options: options} |> Httpc.dispatch()
+      {_, resp} = %Http{url: url, options: options} |> Http.dispatch()
       assert resp.body == "getting a response"
     end
 
@@ -69,7 +66,7 @@ defmodule Hui.Http.HttpcTest do
       end)
 
       # httpc outputs charlist equivalent of "Invalid option {non_existing_http_option,binary} ignored \n"
-      assert capture_log(fn -> %Http{url: url, options: options} |> Httpc.dispatch() end) =~
+      assert capture_log(fn -> %Http{url: url, options: options} |> Http.dispatch() end) =~
                "[73, 110, 118, 97, 108, 105, 100, 32, 111, 112, 116, 105, 111, 110, 32, [123, ['non_existing_option', 44, 'binary'], 125], 32, 105, 103, 110, 111, 114, 101, 100, 32, 10]"
     end
   end
@@ -90,7 +87,7 @@ defmodule Hui.Http.HttpcTest do
           body: "{\"doc\":\"request body\"}",
           headers: [{"content-type", "application/json"}]
         }
-        |> Httpc.dispatch()
+        |> Http.dispatch()
 
       assert 200 = resp.status
     end
@@ -110,7 +107,7 @@ defmodule Hui.Http.HttpcTest do
           body: "{\"doc\":\"request body\"}",
           headers: [{"content-type", "application/json"}]
         }
-        |> Httpc.dispatch()
+        |> Http.dispatch()
 
       assert resp.body == json_resp
     end
@@ -131,7 +128,7 @@ defmodule Hui.Http.HttpcTest do
           headers: [{"content-type", "application/json"}],
           options: options
         }
-        |> Httpc.dispatch()
+        |> Http.dispatch()
 
       assert resp.body == "getting a response"
     end
